hotfix: rename ServiceCard and ServiceCardList to GoodsCardMainPage and GoodCardListMainPage
