+++ src/modules/mainPage/goodsCard/index.jsx
@@ -1,10 +1,10 @@
 import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import classnames from 'classnames';
-import classes from './ServiceCard.module.scss';
+import classes from './style.module.scss';
 import toysImg from '../../../assets/images/images/services-card/toys.png';
 
-const ServiceCard = ({ link, color, title, img, disable }) => {
+const GoodsCardMainPage = ({ link, color, title, img, disable }) => {
   const getCardServiceBackground = classnames(
     classes[`card__service-background_${color}`],
     classes['card__service-background'],
@@ -26,7 +26,7 @@ const ServiceCard = ({ link, color, title, img, disable }) => {
   );
 };
 
-ServiceCard.propTypes = {
+GoodsCardMainPage.propTypes = {
   link: PropTypes.string,
   color: PropTypes.string,
   title: PropTypes.string,
@@ -34,7 +34,7 @@ ServiceCard.propTypes = {
   disable: PropTypes.bool,
 };
 
-ServiceCard.defaultProps = {
+GoodsCardMainPage.defaultProps = {
   link: '/',
   color: 'blue',
   title: 'Игрушки',
@@ -42,4 +42,4 @@ ServiceCard.defaultProps = {
   disable: false,
 };
 
-export default ServiceCard;
+export default GoodsCardMainPage;
