+++ src/modules/mainPage/goodsCard/style.module.scss
