+++ src/modules/mainPage/goodsCardList/style.module.scss
@@ -1,4 +1,4 @@
-.services {
+.goods {
     max-width: 1240px;
     column-gap: 20px;
     padding: 120px 0;
