+++ src/modules/mainPage/specialistCard/style.module.scss
