+++ src/modules/mainPage/specialistCardList/style.module.scss
